+++ spec/unit/active_rel/property_spec.rb
@@ -15,10 +15,6 @@ describe Neo4j::ActiveRel::Property do
         'Clazz'
       end
 
-      def self.uses_classname?
-        false
-      end
-
       include Neo4j::ActiveRel::Property
       include Neo4j::ActiveRel::Types
     end