+++ CHANGELOG.md
@@ -3,7 +3,10 @@ All notable changes to this project will be documented in this file.
 This file should follow the standards specified on [http://keepachangelog.com/]
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleaed]
+## [Unreleased]
+
+### Changed
+- `_classname` property has been completely removed, officially dropping support for Neo4j < 2.1.5.
 
 ### Added